data_generator/requirements.txt
python-dotenv
pymysql
sqlalchemy
mysql-connector-python
faker